Updated to make it work in 2025!
It uses pytubefix instead of pytube, as pytube does not look like it is being maintained.

Converted it to run on mac (used to run on Windows). If you plan to run it on Windows, make sure the paths work correctly. You can look at the previous version to help with this.

If no album art is found, it will now use the youtube video thumbnail instead.