Updated to make it work in 2025!
Found a bug where the initial download of the file (for a single song video) appeared to be twice as long when playing in iTunes or mac. Found the length of the audio on other tags so picked the longest one and created a new file (over the original) with the correct length.

Also fixed some cosmetic changes and fixed the downloading of the watermark pic for videos with multiple soundtracks.